pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mortgage-risk-retention-platform"
version = "0.1.0"
description = "Mortgage Risk & Retention Analytics Platform"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Sheron Schley"}]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic>=2.8.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "streamlit>=1.37.0",
  "requests>=2.32.0",
  "matplotlib>=3.9.0",
  "seaborn>=0.13.0",
  "fpdf2>=2.7.9",
  "python-dotenv>=1.0.1",
  "joblib>=1.4.2",
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["app*", "pipelines*", "dashboard*"]
